int -> uint64_t

+++ trunk/SerializableObjects/SerializableObject.h
@@ -7,7 +7,7 @@ class SerializableObject
     public:
         virtual ~SerializableObject();
         virtual void operator=(const SerializableObject&) = 0;
-        virtual int serialize(void* destinationBuffer) = 0;
+        virtual uint64_t serialize(void* destinationBuffer) = 0;
         virtual void deserialize(void* bufferToUse) = 0;
 };
